Avoid restating conversion types.
