+++ src/pre.ts
@@ -1,16 +1,14 @@
 import { createBoard, BoardUI } from "./board/ui";
 import { FileSystem } from "./board/fs";
 import * as constants from "./board/constants";
+import * as conversions from "./board/conversions";
 
 declare global {
   interface Window {
     board: BoardUI;
     fs: FileSystem;
     constants: typeof constants;
-
-    conversions: {
-      convertAudioBuffer: (source: number, target: AudioBuffer) => void;
-    };
+    conversions: typeof conversions;
 
     HEAPU8: Uint8Array;
   }
@@ -28,14 +26,4 @@ const onSensorChange = () =>
   );
 window.board = createBoard(onSensorChange);
 window.constants = constants;
-window.conversions = {
-  convertAudioBuffer: (source: number, target: AudioBuffer) => {
-    const channel = target.getChannelData(0);
-    const heap = window.HEAPU8;
-    for (let i = 0; i < channel.length; ++i) {
-      // Convert from uint8 to -1..+1 float.
-      channel[i] = (heap[source + i] / 255) * 2 - 1;
-    }
-    return target;
-  },
-};
+window.conversions = conversions;
